moved in-memory-data service to better location, and fixed any import statements that point to it.
