--- client-side/src/app/portals/Caliber/settings/screening/services/in-memory-data.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { InMemoryDataService } from './in-memory-data.service';
-
-describe('InMemoryDataService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [InMemoryDataService]
-    });
-  });
-
-  it('should be created', inject([InMemoryDataService], (service: InMemoryDataService) => {
-    expect(service).toBeTruthy();
-  }));
-});
